FirestoreClass.kt, fun Products from Firebase.

+++ app/src/main/java/com/example/restaurantapp/ui/models/CartItem.kt
@@ -8,7 +8,7 @@ data class CartItem(
     val user_id: String = "",
     val product_id: String = "",
     val title: String = "",
-    val price: Double = 0.0,
+    val price: Long = 0,
     var cart_quantity: Int = 1,
     var android: String = "",
     var id: String = ""
